sr5/requirements.txt
git+ssh://git@github.com/seatgeek/fuzzywuzzy.git@0.15.1#egg=fuzzywuzzy
pyparsing
pint
autopep8
google-api-python-client
sekizai
python-dateutil
python-Levenshtein
